Extract the pivot and unit-conversion blocks out of object_transformer
Both were self-contained subsystems reached from three call sites and calling
nothing else in the class. object_transformer.py drops from 1558 to 1246 lines.

The pivot group (MELT/UNMELT, 5 functions) moves to transformer/pivot.py. Only
two of them touched instance state, and only target_schemaview, which they now
take explicitly.

_perform_unit_conversion moves to functions/unit_conversion.py, next to
convert_units and UnitSystem, which it already depended on. It referenced no
instance state at all — it took everything through DerivationContext — so it
takes the three fields it used (source row, schemaview, source type) as
parameters instead. Depending on DerivationContext would have been circular,
since object_transformer imports this module.

No behaviour change: 1102 passed, 4 skipped, 1 xfailed, unchanged.

Closes #304
